+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Security in case of vulnerabilities.
 - ProjectGroup is created when a project is edited with a new GRN
 - Added `Add, Edit and Delete External Contacts for both conversion and transfer projects.
 - Added `Confirm this transfer has authority to proceed` task
+- Added `Confirm the date the academy transferred` task
+
 
 ### Changed
 - Group reference number links to the group on "About the project"